requirements.txt
fastapi[standard]
pytest
playwright
pytest-playwright
pytest-html
PyYAML
